Fixing a bug which leads to a panic when the key isn't available at all in consul/etc
During testing I experienced that vip-manager is failing totally if the key is not available and also set the status variable explicitly to false which could lead to an ip change. I think in the occasion that something goes wrong and consul/etcd is not reachable it would be nice if vip-manager just does nothing and is also able to revive from such an issue especially because patroni, which this tool most likely is used for, deletes the leader variable during a failover which leads to a failing vip-manager during a patroni failover.