Removed the arpqueryduplicates check
We investigated the issue that a virtual ip get's moved from one server to another while the first server keeps it's virtual ip (malfunctioning). The new server assigns itself the new address and sends a gratuitous arp which will direct every new traffic to that server. Somehow if the second server unassigns the virtual ip again it can be possible that other servers will fallback to the first server again (also with no gratuitous arp) because the arp table cache is 60 seconds and will keep former assignments this long.
After we figured this out, we decided that even a malfunctioning vip-manager which keeps the ip address cannot prevent that another server will take over. After something like this has happened, an engineer has to look what happened and can also take down the server with the same ip address hours later.